+++ Source/Phone/Phone.swift
@@ -306,7 +306,7 @@ public class Phone {
                             if let device = self.devices.device {
                                 let media = MediaModel(sdp: localSDP, audioMuted: false, videoMuted: false, reachabilities: reachabilities)
                                 if target.isEndpoint {
-                                    self.client.create(target.address, by: device, localMedia: media, queue: self.queue.underlying) { res in
+                                    self.client.create(target.address, by: device, localMedia: media, layout: option.layout, queue: self.queue.underlying) { res in
                                         self.doLocusResponse(LocusResult.call(target.isGroup, device, option.uuid, tempMediaContext, res, completionHandler))
                                         self.queue.yield()
                                     }
@@ -314,7 +314,7 @@ public class Phone {
                                 else {
                                     self.conversations.getLocusUrl(conversation: target.address, by: device, queue: self.queue.underlying) { res in
                                         if let url = res.result.data?.locusUrl {
-                                            self.client.join(url, by: device, localMedia: media, queue: self.queue.underlying) { resNew in
+                                            self.client.join(url, by: device, localMedia: media, layout: option.layout, queue: self.queue.underlying) { resNew in
                                                 self.doLocusResponse(LocusResult.call(target.isGroup, device, option.uuid, tempMediaContext, resNew, completionHandler))
                                                 self.queue.yield()
                                             }
@@ -471,7 +471,7 @@ public class Phone {
                 tempMediaContext.prepare(option: option, phone: self)
                 let media = MediaModel(sdp: tempMediaContext.getLocalSdp(), audioMuted: false, videoMuted: false, reachabilities: self.reachability.feedback?.reachabilities)
                 self.queue.sync {
-                    self.client.join(call.url, by: call.device, localMedia: media, queue: self.queue.underlying) { res in
+                    self.client.join(call.url, by: call.device, localMedia: media, layout: option.layout, queue: self.queue.underlying) { res in
                         self.doLocusResponse(LocusResult.join(call, res, completionHandler))
                         self.queue.yield()
                     }